food_vision_playground/requirements.txt
numpy
torch
torchvision
timm==0.6.13
pillow
tqdm
